Fixed some issues concerning media attachments (continue post works)
Dates are working in Win7 on Safari, Chrome, IE, and FF. Note this has only been debugged on discussion.php... expect lots of errors elsewhere.

Also lots of the .php files are changed to set the timezone to UTC time. Going forward we should ensure every time in the DB is stored as UTC; this is then relatively easy to convert to a given user's timezone.